requirements_test.txt
pytest>=7.4
pytest-asyncio>=0.23
pytest-homeassistant-custom-component>=0.13
packaging>=23.0
